Move stuff and small refactor
