+++ src/main/java/net/gensokyoreimagined/nitori/mixin/needs_testing/brain/BrainMixin.java
@@ -1,4 +1,4 @@
-package net.gensokyoreimagined.nitori.mixin.collections.brain;
+package net.gensokyoreimagined.nitori.mixin.needs_testing.brain;
 
 import com.google.common.collect.ImmutableList;
 import it.unimi.dsi.fastutil.objects.Reference2ReferenceLinkedOpenHashMap;
