+++ src/main/java/net/gensokyoreimagined/nitori/mixin/needs_testing/inline_height/WorldChunkMixin.java
@@ -1,4 +1,4 @@
-package net.gensokyoreimagined.nitori.mixin.world.inline_height;
+package net.gensokyoreimagined.nitori.mixin.needs_testing.inline_height;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.LevelHeightAccessor;
