+++ src/main/java/net/gensokyoreimagined/nitori/mixin/needs_testing/inline_height/WorldMixin.java
@@ -1,4 +1,4 @@
-package net.gensokyoreimagined.nitori.mixin.world.inline_height;
+package net.gensokyoreimagined.nitori.mixin.needs_testing.inline_height;
 
 //import net.minecraft.core.RegistryAccess;
 //import net.minecraft.resources.ResourceKey;
