+++ src/main/resources/mixins.core.json
@@ -8,7 +8,6 @@
   "server": [
     "ChunkMapMixin",
     "ChunkMapMixin$TrackedEntity",
-    "math.fast_util.MixinAABB",
     "MixinBlock",
     "MixinBlockPos",
     "MixinChunkEntitySlices",
@@ -19,7 +18,6 @@
     "MixinIteratorSafeOrderedReferenceSet",
     "MixinLevelStorageAccess",
     "MixinMob",
-    "math.sine_lut.MixinMth",
     "MixinNoiseBasedChunkGenerator",
     "MixinPlayer",
     "MixinPlayerList",
@@ -28,49 +26,51 @@
     "MixinServerEntity",
     "MixinSpongeSIMD",
     "MixinWorldGenRegion",
-    "alloc.chunk_ticking.ServerChunkManagerMixin",
     "alloc.blockstate.StateMixin",
-    "alloc.composter.ComposterMixin$ComposterBlockFullComposterInventoryMixin",
-    "alloc.composter.ComposterMixin$ComposterBlockDummyInventoryMixin",
+    "alloc.chunk_ticking.ServerChunkManagerMixin",
     "alloc.composter.ComposterMixin$ComposterBlockComposterInventoryMixin",
-    "util.MixinLevelBlockEntityRetrieval",
+    "alloc.composter.ComposterMixin$ComposterBlockDummyInventoryMixin",
+    "alloc.composter.ComposterMixin$ComposterBlockFullComposterInventoryMixin",
     "cached_hashcode.BlockNeighborGroupMixin",
-    "shapes.blockstate_cache.BlockMixin",
-    "shapes.lazy_shape_context.EntityShapeContextMixin",
-    "entity.fast_retrieval.SectionedEntityCacheMixin",
+    "collections.attributes.AttributeContainerMixin",
+    "collections.block_entity_tickers.WorldChunkMixin",
+    "collections.chunk_tickets.SortedArraySetMixin",
+    "collections.entity_by_type.TypeFilterableListMixin",
+    "collections.entity_filtering.TypeFilterableListMixin",
+    "collections.goals.GoalSelectorMixin",
+    "collections.mob_spawning.SpawnSettingsMixin",
     "entity.fast_hand_swing.LivingEntityMixin",
-    "math.fast_blockops.DirectionMixin",
+    "entity.fast_retrieval.SectionedEntityCacheMixin",
+    "logic.fast_bits_blockpos.OptimizedBlockPosBitsMixin",
     "math.fast_blockops.BlockPosMixin",
-    "math.fast_util.AxisCycleDirectionMixin$ForwardMixin",
+    "math.fast_blockops.DirectionMixin",
     "math.fast_util.AxisCycleDirectionMixin$BackwardMixin",
+    "math.fast_util.AxisCycleDirectionMixin$ForwardMixin",
     "math.fast_util.DirectionMixin",
-    "math.vec.FastMathVec3DMixin",
+    "math.fast_util.MixinAABB",
     "math.general.GenericFastMathMixin",
     "math.intrinsic.MathHelperIntrinsicMixin",
     "math.joml.JOMLMixin",
     "math.rounding.FastRoundingVoxShapeMixin",
-    "collections.entity_filtering.TypeFilterableListMixin",
-    "collections.entity_by_type.TypeFilterableListMixin",
-    "collections.chunk_tickets.SortedArraySetMixin",
-    "collections.block_entity_tickers.WorldChunkMixin",
-    "collections.goals.GoalSelectorMixin",
-    "collections.mob_spawning.SpawnSettingsMixin",
-    "collections.attributes.AttributeContainerMixin",
-    "world.block_entity_ticking.sleeping.WrappedBlockEntityTickInvokerAccessor",
-    "world.block_entity_ticking.sleeping.campfire.CampfireBlockEntityMixin",
-    "world.block_entity_ticking.sleeping.campfire.unlit.CampfireBlockEntityMixin",
-    "world.block_entity_ticking.sleeping.campfire.lit.CampfireBlockEntityMixin",
-    "world.block_entity_ticking.sleeping.furnace.AbstractFurnaceBlockEntityMixin",
-    "world.block_entity_ticking.support_cache.DirectBlockEntityTickInvokerMixin",
-    "world.block_entity_ticking.support_cache.BlockEntityMixin",
-    "world.block_entity_ticking.support_cache.WorldChunkMixin",
+    "math.sine_lut.MixinMth",
+    "math.vec.FastMathVec3DMixin",
+    "playerwatching.MixinChunkFilter",
+    "playerwatching.MixinServerPlayerEntity",
+    "playerwatching.optimize_nearby_player_lookups.MixinMobEntity",
+    "shapes.blockstate_cache.BlockMixin",
+    "shapes.lazy_shape_context.EntityShapeContextMixin",
+    "util.MixinLevelBlockEntityRetrieval",
     "util.accessors.ClientEntityManagerAccessor",
     "util.accessors.EntityTrackingSectionAccessor",
     "util.accessors.ServerEntityManagerAccessor",
     "util.block_tracking.AbstractBlockStateMixin",
-    "playerwatching.MixinChunkFilter",
-    "playerwatching.MixinServerPlayerEntity",
-    "playerwatching.optimize_nearby_player_lookups.MixinMobEntity",
-    "logic.fast_bits_blockpos.OptimizedBlockPosBitsMixin"
+    "world.block_entity_ticking.sleeping.WrappedBlockEntityTickInvokerAccessor",
+    "world.block_entity_ticking.sleeping.campfire.CampfireBlockEntityMixin",
+    "world.block_entity_ticking.sleeping.campfire.lit.CampfireBlockEntityMixin",
+    "world.block_entity_ticking.sleeping.campfire.unlit.CampfireBlockEntityMixin",
+    "world.block_entity_ticking.sleeping.furnace.AbstractFurnaceBlockEntityMixin",
+    "world.block_entity_ticking.support_cache.BlockEntityMixin",
+    "world.block_entity_ticking.support_cache.DirectBlockEntityTickInvokerMixin",
+    "world.block_entity_ticking.support_cache.WorldChunkMixin"
   ]
 }
